Fix performance degradation of HIP dot
The results did not match between cuda-stream and hip-stream on the same
NVIDIA GPU card (NVIDIA A100 40GB PCIe) when large arraysize is specified.
cuda-stream uses the number of SMs to decide dot_num_blocks, which looks
more sensible than to use arraysize to determine the parameter. It is used
as kernel grid size and iteration count for reduction in the host code.

Link: https://github.com/UoB-HPC/BabelStream/commit/9954b7d38cd85d20927428425a9840a72a56c3e4
Signed-off-by: Daisuke Matsuda <matsuda-daisuke@fujitsu.com>